Product Manager Group Therapy content
Here's the first iteration on Product Manager Group Therapy content. I'll likely add more as I continue to work on this presentation, but for now, it's enough to share with my product peeps who attended the March 22, 2019 ProductTank  RTP meetup where I presented this.

Update ProductTherapy.DeanPeters.21Mar19.Final.pdf

Was missing a slide, not it's not.